% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyboots.R
\name{shinyboots}
\alias{shinyboots}
\title{shinyboots}
\usage{
shinyboots()
}
\value{
Histograms of the log(lambda) and Ybar distributions with vertical line indicating the cut-off value for acceptance/rejection regions
}
\description{
An interactive app with a slider for the alpha value
}
\examples{
\dontrun{shinyboots()}
}
