% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinypower.R
\name{shinypower}
\alias{shinypower}
\title{shinypower}
\usage{
shinypower()
}
\value{
Plot of the Power vs the sample size n
}
\description{
An interactive app for the power of a distribution
}
\examples{
\dontrun{shinypower()}
}
